Added Maven support & Logging minor introduction. Some refactoring
