--- src/TextGenerator/parsers/TxtParser.java
@@ -1,9 +1,0 @@
-package TextGenerator.parsers;
-
-import java.io.BufferedReader;
-import java.io.FileReader;
-import java.io.IOException;
-
-public class TxtParser {
-
-}
