+++ src/main/java/TextGenerator/parsers/TxtParser.java
@@ -1,0 +1,9 @@
+package TextGenerator.parsers;
+
+import java.io.BufferedReader;
+import java.io.FileReader;
+import java.io.IOException;
+
+public class TxtParser {
+
+}
